requirements.txt
httpx>=0.28.1
mcp>=1.2.0
wolframalpha>=5.1.3
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
multidict>=6.0.0
xmltodict>=0.13.0 